# Conflicts:
#	CMakeLists.txt
#	Common/core/def_version.h
#	Common/game/interactions.h
#	Common/game/main_game_file.cpp
#	Common/game/room_file.cpp
#	Common/gui/guidefines.h
#	Common/gui/guimain.cpp
#	Common/gui/guimain.h
#	Common/script/cc_script.cpp
#	Common/util/math.h
#	Common/util/string_utils.cpp
#	Editor/AGS.Editor/Resources/agsdefns.sh
#	Editor/AGS.Editor/app.manifest
#	Editor/AGS.Types/Properties/AssemblyInfo.cs
#	Engine/ac/dynobj/cc_character.cpp
#	Engine/ac/dynobj/cc_character.h
#	Engine/ac/dynobj/scriptdrawingsurface.cpp
#	Engine/ac/global_inventoryitem.cpp
#	Engine/ac/global_object.cpp
#	Engine/ac/global_string.cpp
#	Engine/ac/guicontrol.cpp
#	Engine/ac/invwindow.cpp
#	Engine/ac/mouse.cpp
#	Engine/ac/overlay.cpp
#	Engine/ac/room.cpp
#	Engine/ac/roomobject.cpp
#	Engine/ac/roomobject.h
#	Engine/ac/screenoverlay.h
#	Engine/game/game_init.cpp
#	Engine/game/savegame.cpp
#	Engine/game/savegame_components.cpp
#	Engine/game/savegame_v321.cpp
#	Engine/gfx/ali3dsw.h
#	Engine/gfx/gfxfilter_aad3d.cpp
#	Engine/main/game_run.cpp
#	Engine/platform/android/acpland.cpp
#	Engine/script/script.cpp
#	Engine/script/script.h
#	version.json

+++ Common/ac/view.cpp
@@ -82,7 +82,7 @@ void ViewLoopNew::Dispose()
 
 void ViewLoopNew::WriteToFile_v321(Stream *out)
 {
-    out->WriteInt16(numFrames);
+    out->WriteInt16(static_cast<uint16_t>(numFrames));
     out->WriteInt32(flags);
     WriteFrames_Aligned(out);
 }
@@ -99,7 +99,7 @@ void ViewLoopNew::WriteFrames_Aligned(Stream *out)
 
 void ViewLoopNew::ReadFromFile_v321(Stream *in)
 {
-    Initialize(in->ReadInt16());
+    Initialize(static_cast<uint16_t>(in->ReadInt16()));
     flags = in->ReadInt32();
     ReadFrames_Aligned(in);
 }
@@ -133,7 +133,7 @@ void ViewStruct::Dispose()
 
 void ViewStruct::WriteToFile(Stream *out)
 {
-    out->WriteInt16(numLoops);
+    out->WriteInt16(static_cast<uint16_t>(numLoops));
     for (int i = 0; i < numLoops; i++)
     {
         loops[i].WriteToFile_v321(out);
@@ -142,7 +142,7 @@ void ViewStruct::WriteToFile(Stream *out)
 
 void ViewStruct::ReadFromFile(Stream *in)
 {
-    Initialize(in->ReadInt16());
+    Initialize(static_cast<uint16_t>(in->ReadInt16()));
 
     for (int i = 0; i < numLoops; i++)
     {
