+++ Common/ac/view.h
@@ -38,7 +38,7 @@ struct ViewFrame {
 
 struct ViewLoopNew
 {
-    short numFrames;
+    int   numFrames;
     int   flags;
     std::vector<ViewFrame> frames;
     // NOTE: we still need numFrames for backward compatibility:
@@ -57,7 +57,7 @@ struct ViewLoopNew
 
 struct ViewStruct
 {
-    short numLoops;
+    int numLoops;
     std::vector<ViewLoopNew> loops;
 
     ViewStruct();
